requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.2
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.3
django_celery_results==2.6.0
dotenv==0.9.9
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.52
psycopg2==2.9.10
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
redis==7.0.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
text-unidecode==1.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
